User/src/test/java/com/proyect/User/controller/ShoppingControllerTest.java
package com.proyect.User.controller;

import static org.junit.jupiter.api.Assertions.*;

import org.junit.jupiter.api.Test;

class ShoppingControllerTest {

	@Test
	void shouldReturnTrueInValidateCard() {
		ShoppingController sc = new ShoppingController();
		assertTrue(sc.validateCard("1234567891234567", 5, 2023));
	}
	
	@Test
	void shouldReturnFalseInValidateCardByWrongLengthNumberSoShort() {
		ShoppingController sc = new ShoppingController();
		assertFalse(sc.validateCard("123456789", 12, 2022));
	}
	
	@Test
	void shouldReturnFalseInValidateCardByWrongLengthNumberSoLong() {
		ShoppingController sc = new ShoppingController();
		assertFalse(sc.validateCard("123456789123456789", 12, 2022));
	}
	
	@Test
	void shouldReturnFalseInValidateCardByWrongMonthLessThan1() {
		ShoppingController sc = new ShoppingController();
		assertFalse(sc.validateCard("11234567891234567", 0, 2022));
	}
	
	@Test
	void shouldReturnFalseInValidateCardByWrongMonthGreaterThan12() {
		ShoppingController sc = new ShoppingController();
		assertFalse(sc.validateCard("11234567891234567", 13, 2022));
	}
	
	@Test
	void shouldReturnFalseInValidateCardByWrongYearPast() {
		ShoppingController sc = new ShoppingController();
		assertFalse(sc.validateCard("11234567891234567", 6, 2021));
	}
	
	@Test
	void shouldReturnFalseInValidateCardByWrongYearIncrediblyFuture() {
		ShoppingController sc = new ShoppingController();
		assertFalse(sc.validateCard("11234567891234567", 6, 2300));
	}
	
	
	
	

}
